docker_paperless/mkdocs/requirements.txt

python-telegram-bot
requests
pyyaml
